Z1013 Software Datenbank

+++ app_04_z1013-128/tools/add_db.sh
@@ -68,7 +68,9 @@ if (-f $video) {
 }
 
 $gif=$ENV{"HOME"}."/jkcemu.gif";
-unlink($gif) or warn "Could not unlink $gif: $!";
+if (-e $gif) {
+    unlink($gif) or warn "Could not unlink $gif: $!";
+}
 
 @list = bsd_glob ("~/jkcemu_screen_*.txt");
 for (@list) {
